+++ packages/server/hypermedia/template/transform.ts
@@ -19,7 +19,6 @@ export type HtmlTransformVisitFunctions = {
 };
 
 export type HtmlTransformVisitor = (node: HtmlNode, fns: HtmlTransformVisitFunctions) => HtmlTransformVisitResponse;
-
 export type HtmlSimpleTransformVisitor = (node: HtmlNode) => HtmlTransformVisitResponse;
 
 export function transformHtml(node: HtmlNode, visit: HtmlTransformVisitor) {
@@ -51,11 +50,60 @@ export function transformHtml(node: HtmlNode, visit: HtmlTransformVisitor) {
  * @param visit - The visitor function to apply to each HTML node.
  */
 export function simpleTransformHtml(node: HtmlNode, visit: HtmlSimpleTransformVisitor) {
-    transformHtml(node, (node, { visitEachChild, visitNode }) => {
+    transformHtml(node, (node, { visitEachChild }) => {
         const results = visit(node);
         const nodes = [results].flat().filter((it) => it) as HtmlNode[];
         for (const node of nodes) {
             visitEachChild(node);
+        }
+        return nodes;
+    });
+}
+
+export type HtmlAsyncTransformVisitResponse = Promise<HtmlNode | (HtmlNode | undefined)[] | undefined>;
+
+export type HtmlAsyncTransformVisitNodeFunction = (node: HtmlNode) => HtmlAsyncTransformVisitResponse;
+export type HtmlAsyncTransformVisitEachChildFunction = (node: HtmlNode) => Promise<HtmlNode>;
+
+export type HtmlAsyncTransformVisitFunctions = {
+    visitNode: HtmlAsyncTransformVisitNodeFunction;
+    visitEachChild: HtmlAsyncTransformVisitEachChildFunction;
+};
+
+export type HtmlAsyncTransformVisitor = (
+    node: HtmlNode,
+    fns: HtmlAsyncTransformVisitFunctions,
+) => HtmlAsyncTransformVisitResponse;
+export type HtmlSimpleAsyncTransformVisitor = (node: HtmlNode) => HtmlAsyncTransformVisitResponse;
+
+export async function asyncTransformHtml(node: HtmlNode, visit: HtmlAsyncTransformVisitor) {
+    async function visitNode(node: HtmlNode) {
+        return await visit(node, { visitNode, visitEachChild });
+    }
+
+    async function visitEachChild(node: HtmlNode) {
+        if (node.type === "fragment" || node.type === "element") {
+            const replacements = [];
+            for (const child of node.children) {
+                replacements.push(await visit(child, { visitEachChild, visitNode }));
+            }
+            node.children = replacements.flat().filter((it) => it) as HtmlNode[];
+            for (const child of node.children) {
+                child.parent = node;
+            }
+        }
+        return node;
+    }
+
+    return await visitNode(node);
+}
+
+export async function simpleAsyncTransformHtml(node: HtmlNode, visit: HtmlSimpleAsyncTransformVisitor) {
+    return await asyncTransformHtml(node, async (node, { visitEachChild }) => {
+        const results = await visit(node);
+        const nodes = [results].flat().filter((it) => it) as HtmlNode[];
+        for (const node of nodes) {
+            await visitEachChild(node);
         }
         return nodes;
     });